add newproduct function

+++ src/main/java/com/example/FarmMarket/service/FarmMarketService.java
@@ -1,10 +1,11 @@
 package com.example.FarmMarket.service;
 
+import com.example.FarmMarket.objects.PopUpWindow;
 import com.example.FarmMarket.repository.FarmMarketRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Service
 public class FarmMarketService {
@@ -16,7 +17,7 @@ public class FarmMarketService {
     }
 
     public void newProduct(String category, String product_name, String product_description,
-                           BigInteger price, BigInteger amount) {
+                                  BigDecimal price, BigDecimal amount) {
         farmMarketRepository.newProduct(category, product_name, product_description, price, amount);
     }
